Success in using _alltoallw at last. Declaration of MPI_Aint type on the Lua side was initially wrong (unsigned, but somehow worked fine with MPISHM? no way of knowing as there are no MPISH unit tests), so changing it to ptrdiff_t fixes the _alltoallw seg fault and results appear correct.

+++ Comm/Mpi.lua
@@ -36,7 +36,7 @@ ffi.cdef [[
   typedef struct MPI_Request_type *MPI_Request;
   typedef struct MPI_Info_type *MPI_Info;
   typedef struct MPI_Win_type *MPI_Win;
-  typedef unsigned MPI_Aint; /* Not sure if this is correct, but it seems to work */
+  typedef ptrdiff_t MPI_Aint; /* Not sure if this is correct, but it seems to work */
 
   // size of various objects
   int sizeof_MPI_Status();
